# internal/swag/swagger.yaml
basePath: /
definitions:
  handler.CreateVehicleRequest:
    type: object
  handler.ErrorResponse:
    properties:
      detail:
        type: string
      error:
        type: string
    type: object
...
  handler.UpdateProfileRequest:
    properties:
      avatar_url:
        type: string
      display_name:
        maxLength: 100
        minLength: 1
        type: string
    type: object
  handler.UpdateVehicleRequest:
    type: object
  handler.UserProfileResponse:
    properties:
      avatar_url:
        type: string
      created_at:
        type: string
      display_name:
        type: string
      email:
        type: string
      id:
        type: string
      updated_at:
        type: string
      username:
        type: string
    type: object
  handler.VehicleResponse:
    properties:
      brand:
        type: string
      color:
        type: string
      created_at:
        type: string
      id:
        type: string
      is_active:
        type: boolean
      name:
        type: string
      type:
        type: string
      updated_at:
        type: string
    type: object
host: api.trackride.app
info:
  contact:
    name: TrackRide Support
    url: https://github.com/nashirabbash/trackride
  description: Backend API for the TrackRide mobile application. Provides user authentication,
    vehicle tracking, ride logging, and social features.
  license:
    name: MIT
    url: https://github.com/nashirabbash/trackride/blob/master/LICENSE.txt
  title: TrackRide API
  version: 1.0.0
paths:
...
  /v1/users/{id}:
    get:
      description: Retrieve a public user profile. Does not return email address.
      parameters:
      - description: User ID (UUID format)
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: object
        "400":
          description: Invalid ID format
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "404":
          description: User not found
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get user profile by ID
      tags:
      - Users
...
  /v1/users/me:
    get:
      description: Retrieve the authenticated user's complete profile
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.UserProfileResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "404":
          description: User not found
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get current user profile
      tags:
      - Users
    put:
      consumes:
      - application/json
      description: Update the authenticated user's display name and/or avatar URL
      parameters:
      - description: Profile update data (all fields optional)
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/handler.UpdateProfileRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.UserProfileResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "404":
          description: User not found
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "422":
          description: Validation error
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Update current user profile
      tags:
      - Users
  /v1/vehicles:
    get:
      description: Retrieve all vehicles owned by the authenticated user
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/handler.VehicleResponse'
            type: array
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
      security:
      - BearerAuth: []
      summary: List user vehicles
      tags:
      - Vehicles
    post:
      consumes:
      - application/json
      description: 'Create a new vehicle for the authenticated user. Type must be
        one of: motor, mobil, sepeda'
      parameters:
      - description: Vehicle data
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/handler.CreateVehicleRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/handler.VehicleResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "422":
          description: Validation error
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Create new vehicle
      tags:
      - Vehicles
  /v1/vehicles/{id}:
    delete:
      description: Delete a vehicle owned by the authenticated user. Cannot delete
        if vehicle is in use in active rides.
      parameters:
      - description: Vehicle ID (UUID format)
        in: path
        name: id
        required: true
        type: string
      responses:
        "204":
          description: No Content
        "400":
          description: Invalid ID format
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "404":
          description: Vehicle not found
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "409":
          description: Vehicle is in use
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Delete vehicle
      tags:
      - Vehicles
    put:
      consumes:
      - application/json
      description: Update a vehicle owned by the authenticated user. All fields are
        optional.
      parameters:
      - description: Vehicle ID (UUID format)
        in: path
        name: id
        required: true
        type: string
      - description: Vehicle update data
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/handler.UpdateVehicleRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.VehicleResponse'
        "400":
          description: Invalid ID format
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "404":
          description: Vehicle not found
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "422":
          description: Validation error
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Update vehicle
      tags:
      - Vehicles
schemes:
- https
securityDefinitions:
  BearerAuth:
    description: 'JWT Bearer token authentication. Pass as "Authorization: Bearer
      <token>"'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"